project-coincap-streaming-pipeline/requirements.txt
python-dotenv==1.0.1
websocket-client==1.6.1
kafka-python==1.4.7
google-cloud-storage==2.14.0
pytest
confluent-kafka==2.3.0